+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.3.0.0
 * Use new key sharing scheme: Different keys used in different directions; keys rotated; secret keys retrieved via RSA public key rather than embedding in the URL
 * [bugfix] Terminal dimensions in browser match upon initial connection, instead of after resizing
+* Allow static site to route terminal traffic through other server. If static site is detected, user can enter the terminal id and server url in the browser UI.
+
 
 ## 0.2.0.0
 * Add ability to copy+paste using keystrokes (copy with ctrl+shift+c or ctrl+shift+x, and paste with ctrl+shift+v)